Add unit test for HTTP/2 Plain Text
Move HTTP/2 tests to separate test class.
Move helper method for creating client to Utils.
